Resolucion ejercicio

+++ Practica_Calificada_01/Ejercicio02.js
@@ -1,5 +1,5 @@
 let edad= parseInt(prompt("ingrese su edad"));
-let ingreso= parseInt(prompt("ingrese su ingreso mensual"));
+let ingreso= parseFloat(prompt("ingrese su ingreso mensual"));
 
 if (edad>=18 && ingreso>=500 && ingreso <= 2000){
     console.log("el usuario es legible para el subsidio")
